move files to better places, delete PlayerRig
