+++ Assets/Scripts/OperatorUserInterface/EmotionManager.cs
