+++ Assets/Scripts/OperatorUserInterface/EmotionManager.cs.meta
