--- Assets/Scripts/PlayerRig.cs
@@ -1,7 +1,0 @@
-﻿/// <summary>
-/// This script is only used to get the reference to this GameObject from any scene.
-/// </summary>
-public class PlayerRig : Singleton<PlayerRig>
-{
-    
-}
