--- Assets/Scripts/PlayerRig.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a4f81dd52bce6c54aad5d8db5c7666d9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
